Change to 'type Investment' on files that imports Investment interface. Changed also investment-result's location and name, it changed to service. Add one more interface named AnnualData to investment.model.
