+++ src/app/investment.model.ts
@@ -4,3 +4,12 @@ export interface Investment{
   expectedReturn: number,
   duration: number
 }
+
+export interface AnnualData{
+  year: number,
+  interest: number,
+  valueEndOfYear: number,
+  annualInvestment: number,
+  totalInterest: number,
+  totalAmountInvested: number
+}
